LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/document-access/document_access_parent_access.tentative-expected.txt

CONSOLE MESSAGE: line 2: TypeError: Not enough arguments


FAIL Untitled assert_equals: Security Error thrown expected "SecurityError" but got "ReferenceError"
